+++ shared/gh/js/utils/gh.utils.time.js
@@ -1,0 +1,462 @@
+/*!
+ * Copyright 2014 Digital Services, University of Cambridge Licensed
+ * under the Educational Community License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at
+ *
+ *     http://opensource.org/licenses/ECL-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an "AS IS"
+ * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+define(['exports', 'gh.constants', 'moment'], function(exports, constants, moment) {
+
+    /**
+     * Add a leading zero to a digit
+     *
+     * @param  {Number}    digit    The digit that needs to be extended with an extra zero, if necessary
+     * @return {Number}             The extended digit
+     */
+    var addLeadingZero = exports.addLeadingZero = function(digit) {
+        if (!_.isNumber(digit)) {
+            throw new Error('An invalid digit has been provided');
+        }
+
+        // Convert the digit to a string
+        digit = String(digit);
+
+        // Add a leading zero if the length of the string equals 1
+        if (digit.length === 1) {
+            digit = '0' + digit;
+        }
+        return digit;
+    };
+
+    /**
+     * Convert an ISO8601 date to a UNIX date
+     *
+     * @param  {String}    date    The ISO8601 date that needs to be converted to a UNIX date format
+     * @return {Number}            The UNIX date
+     */
+    var convertISODatetoUnixDate = exports.convertISODatetoUnixDate = function(date) {
+        if (!date || !_.isString(date) || !moment(date, 'YYYY-MM-DD').isValid()) {
+            throw new Error('An invalid value for date has been provided');
+        }
+        return Date.parse(date);
+    };
+
+    /**
+     * Convert a UNIX date to an ISO8601 date
+     *
+     * @param  {String}    date    The UNIX date that needs to be converted to an ISO8601 date format
+     * @return {Number}            The ISO8601 date
+     */
+    var convertUnixDatetoISODate = exports.convertUnixDatetoISODate = function(date) {
+        if (!date || !moment(date, 'x').isValid()) {
+            throw new Error('An invalid value for date has been provided');
+        }
+        return new Date(date).toISOString();
+    };
+
+    /**
+     * @param  {String}    startDate    The start date of the event in UTC format (e.g. 2014-10-07T08:00:00.000Z)
+     * @param  {String}    endDate      The end date of the event in UTC format (e.g. 2014-10-07T11:00:00.000Z)
+     * @return {String}                 The converted date in display format (e.g. 'W7 • Fri 2–3pm')
+     */
+    var generateDisplayDate = exports.generateDisplayDate = function(startDate, endDate) {
+        if (!_.isString(startDate) || !moment(startDate, 'YYYY-MM-DD').isValid()) {
+            throw new Error('A valid start date should be provided');
+        } else if (!_.isString(endDate) || !moment(endDate, 'YYYY-MM-DD').isValid()) {
+            throw new Error('A valid end date should be provided');
+        }
+
+        // TODO: add the week number when the custom configuration is in place
+        // var weekNumber = 'W' + getWeekInTerm(startDate);
+        var weekNumber = getWeekInTerm(startDate);
+        if (weekNumber === 0) {
+            weekNumber = 'OT';
+        } else {
+            weekNumber = 'W' + weekNumber;
+        }
+
+        // Retrieve the day
+        var weekDay = moment(endDate).utc().format('E');
+        weekDay = constants.time.DAYS[weekDay - 1];
+
+        // Retrieve the meridien values
+        var startMeridien = moment(startDate).utc().format('a');
+        var endMeridien = moment(endDate).utc().format('a');
+
+        // Compose the start part
+        var startParts = [];
+        startParts.push(moment(startDate).utc().format('h'));
+        var startMinutes = moment(startDate).utc().format('mm');
+        if (parseInt(startMinutes, 10)) {
+            startParts.push(':' + startMinutes);
+        }
+        if (startMeridien !== endMeridien) {
+            startParts.push(startMeridien);
+        }
+        startParts = startParts.join('');
+
+        // Compose the end part
+        var endParts = [];
+        endParts.push(moment(endDate).utc().format('h'));
+        var endMinutes = moment(endDate).utc().format('mm');
+        if (parseInt(endMinutes, 10)) {
+            endParts.push(':' + endMinutes);
+        }
+        endParts.push(endMeridien);
+        endParts = endParts.join('');
+
+        // Return the display date
+        return String(weekNumber + ' · ' + weekDay + ' ' + startParts + '-' + endParts);
+    };
+
+    /**
+     * Get the number of weeks in a term
+     *
+     * @param  {Object}    term    The term to get the number of weeks for
+     * @return {Number}            The number of weeks in the term
+     */
+    var getWeeksInTerm = exports.getWeeksInTerm = function(term) {
+        if (!_.isObject(term)) {
+            throw new Error('A valid term should be provided');
+        }
+
+        // The number of milliseconds in one week
+        var ONE_WEEK = 1000 * 60 * 60 * 24 * 7;
+        // Convert the term start and end date to milliseconds
+        var termStartDate = new Date(term.start).getTime();
+        var termEndDate = new Date(term.end).getTime();
+        // Calculate the time difference
+        var timeDifference = Math.abs(termEndDate - termStartDate);
+        // Convert to weeks and return
+        return Math.floor(timeDifference / ONE_WEEK) + 1;
+    };
+
+    /**
+     * Get the Date of the first day of the specified term
+     *
+     * @param  {String}    termName    The name of the term to get the date of the first day for
+     * @return {Date}                  The date of the first day of the term
+     */
+    var getFirstDayOfTerm = exports.getFirstDayOfTerm = function(termName) {
+        if (!_.isString(termName)) {
+            throw new Error('A valid term name should be provided');
+        }
+
+        // Get the configuration
+        var config = require('gh.core').config;
+        // Get the correct terms associated to the current application
+        var terms = config.terms[config.academicYear];
+        // Find the term to get the first day's date for
+        var term =_.find(terms, function(term) {
+            /* istanbul ignore else */
+            if (term.name === termName) {
+                // Return the term
+                return term;
+            }
+        });
+
+        // Return the first day's date
+        return new Date(term.start);
+    };
+
+    /**
+     * Get a date by specifying the term it's in, the week number it's in and the day of the week it is
+     *
+     * @param  {String}    termName      The name of the term to look for the date
+     * @param  {Number}    weekNumber    The week of the term to look for the date
+     * @param  {Number}    dayNumber     The day of the week to look for the dae
+     * @return {Date}                    Date object of the day in the term
+     */
+    var getDateByWeekAndDay = exports.getDateByWeekAndDay = function(termName, weekNumber, dayNumber) {
+        if (!_.isString(termName)) {
+            throw new Error('A valid term name should be provided');
+        } else if (!_.isNumber(weekNumber)) {
+            throw new Error('A valid week number should be provided');
+        } else if (!_.isNumber(dayNumber)) {
+            throw new Error('A valid day number should be provided');
+        }
+
+        // Get the configuration
+        var config = require('gh.core').config;
+        // Get the correct terms associated to the current application
+        var terms = config.terms[config.academicYear];
+        // Variable used to assign the date by week and day to and return
+        var dateByWeekAndDay = null;
+
+        // Loop over the terms
+        _.each(terms, function(term) {
+            if (term.name === termName) {
+                // The number of milliseconds in one week
+                var ONE_WEEK = 1000 * 60 * 60 * 24 * 7;
+                // The number of milliseconds in one day
+                var ONE_DAY = 1000 * 60 * 60 * 24;
+                // Get the date on which the term starts
+                var termStartDate = new Date(term.start).getTime();
+
+                // Calculate the week offset in milliseconds
+                var weekOffset = weekNumber * ONE_WEEK;
+                // Calculate the start date of the week
+                weekOffset = termStartDate + weekOffset;
+
+                // Now calculate the day offset in this week
+                var dayOffset = dayNumber - new Date(weekOffset).getDay();
+                // Calculate the week offset in milliseconds
+                dayOffset = dayOffset * ONE_DAY;
+
+                // Calculate the day of the week to return
+                dateByWeekAndDay = new Date(weekOffset + dayOffset);
+            }
+        });
+
+        // return the Date object
+        return dateByWeekAndDay;
+    };
+
+    /**
+     * Get the week of the term in which a date is located. The function assumes that the
+     * week starts on the first day of the term and that the terms are limited by the
+     * academicYear that is set on the app
+     *
+     * @param  {String|Date}    date    The date to find the week number in the term for
+     * @return {Number}                 The week number of the term the given date is in
+     * @private
+     */
+    var getWeekInTerm = exports.getWeekInTerm = function(date) {
+        // Get the configuration
+        var config = require('gh.core').config;
+        // Get the correct terms associated to the current application
+        var terms = config.terms[config.academicYear];
+        // Variable used to assign the week number to and return
+        var weekNumber = 0;
+
+        // Loop over the terms. If the start date of the event falls inside of the term,
+        // use that term to calculate the week number that date falls in
+        _.each(terms, function(term) {
+            var termStartDate = new Date(term.start);
+            var termEndDate = new Date(term.end);
+            date = new Date(date);
+            // If the date falls in the term, use it to calculate the week number
+            if (termStartDate <= date && termEndDate >= date) {
+                // The number of milliseconds in one week
+                var ONE_WEEK = 1000 * 60 * 60 * 24 * 7;
+                // Convert the given date to milliseconds
+                date = date.getTime();
+                // Convert the term start date to milliseconds
+                termStartDate = termStartDate.getTime();
+                // Calculate the difference in milliseconds
+                var dateDifference = Math.abs(date - termStartDate);
+                // Convert back to weeks
+                weekNumber = Math.floor(dateDifference / ONE_WEEK) + 1;
+            }
+        });
+
+        return weekNumber;
+    };
+
+    /**
+     * All the functionality related to date displaying
+     *
+     * @param  {String}    date    The date in UTC format
+     * @return {Object}            Object containing dateDisplay functions
+     */
+    var dateDisplay = exports.dateDisplay = function(date) {
+
+        /**
+         * Returns the day from a UTC string
+         *
+         * @return {Number}
+         */
+        var dayNumber = function() {
+            if (!_.isString(date) || !moment(date, 'YYYY-MM-DD').isValid()) {
+                throw new Error('A valid date should be provided');
+            }
+            return parseInt(moment(date).utc().format('D'), 10);
+        };
+
+        /**
+         * Returns the month from a UTC tring
+         *
+         * @return {String}
+         */
+        var monthName = function() {
+            if (!_.isString(date) || !moment(date, 'YYYY-MM-DD').isValid()) {
+                throw new Error('A valid date should be provided');
+            }
+            return constants.time.MONTHS[moment(date).utc().format('M') - 1];
+        };
+
+        return {
+            'dayNumber': dayNumber,
+            'monthName': monthName
+        };
+    };
+
+    /**
+     * Convert a date to GMT+0 for display in the calendar
+     *
+     * @param  {String}    date    The date that needs conversion
+     */
+    var fixDateToGMT = exports.fixDateToGMT = function(date) {
+        if (!_.isString(date) || !moment(date, 'YYYY-MM-DD').isValid()) {
+            throw new Error('A valid date should be provided');
+        }
+        return (new Date(date)).getTime() - ((new Date(date)).getTimezoneOffset() * 60000);
+    };
+
+    /**
+     * Convert start and end times of an event to GMT+0 for display in the calendar
+     *
+     * @param  {Object[]}    events    An Array of events to fix start and end date to GTM+0 for
+     */
+    var fixDatesToGMT = exports.fixDatesToGMT = function(events) {
+        if (!_.isArray(events)) {
+            throw new Error('A valid array should be provided');
+        }
+        _.each(events, function(ev) {
+            ev.start = fixDateToGMT(ev.start);
+            ev.end = fixDateToGMT(ev.end);
+        });
+    };
+
+    /**
+     * Get the date range the calendar should be displaying. The date is determined by the calendar's current view.
+     *
+     *  * Day:
+     *  * We only fetch the events that occur in a 24 hour time frame based on the current day
+     *
+     *  * Week:
+     *  * We fetch the events that occur one week before and after the current view's date
+     *
+     *  * Month:
+     *  * We fetch the events that occur one month before and after the current view's date
+     *
+     * @param  {Function}   callback            Standard callback function
+     * @param  {Object}     callback.range      Object containg start and end date that form the range of the calendar
+     */
+    /* istanbul ignore next */
+    var getCalendarDateRange = exports.getCalendarDateRange = function(callback) {
+
+        // Create the range object to return
+        var range = {
+            'start': null,
+            'end': null
+        };
+
+        /**
+         * TODO: move the dispatching of events to actual UI files
+         */
+
+        // Fetch the calendar's current view type
+        $(document).trigger('gh.calendar.getCurrentView', function(currentView) {
+
+            // Fetch the calendar's current view date
+            $(document).trigger('gh.calendar.getCurrentViewDate', function(currentViewDate) {
+
+                // Calculate the start date
+                range.start = convertUnixDatetoISODate(currentViewDate - constants.time.PERIODS[currentView]);
+
+                // Calculate the end date
+                range.end = convertUnixDatetoISODate(currentViewDate + constants.time.PERIODS[currentView]);
+
+                // Return the range object
+                return callback(range);
+            });
+        });
+    };
+
+    /**
+     * Determine whether or not a given date is in the range of 2 dates
+     *
+     * @param  {Number}    date         The date in UNIX format
+     * @param  {Number}    startDate    The start of the date range in UNIX format
+     * @param  {Number}    endDate      The end of the date range in UNIX format
+     * @return {Boolean}                Whether or not the date is in the range
+     */
+    var isDateInRange = exports.isDateInRange = function(date, startDate, endDate) {
+        if (!date || !moment(date, 'x').isValid()) {
+            throw new Error('An invalid value for date has been provided');
+        } else if (!startDate || !moment(startDate, 'x').isValid()) {
+            throw new Error('An invalid value for startDate has been provided');
+        } else if (!endDate || !moment(endDate, 'x').isValid()) {
+            throw new Error('An invalid value for endDate has been provided');
+        } else if (startDate > endDate) {
+            throw new Error('The startDate cannot be after the endDate');
+        }
+
+        return (date >= startDate && date <= endDate);
+    };
+
+    /**
+     * Return the number of weeks within a date range
+     *
+     * @param  {Number}    startDate    The start of the date range in UNIX format
+     * @param  {Number}    endDate      The end of the date range in UNIX format
+     * @return {Number}                 The number of weeks within the date range
+     */
+    var weeksInDateRange = exports.weeksInDateRange = function(startDate, endDate) {
+        if (!startDate || !moment(startDate, 'x').isValid()) {
+            throw new Error('An invalid value for startDate has been provided');
+        } else if (!endDate || !moment(endDate, 'x').isValid()) {
+            throw new Error('An invalid value for endDate has been provided');
+        } else if (startDate > endDate) {
+            throw new Error('The startDate cannot be after the endDate');
+        }
+
+        // Calculate the difference between the two dates and return the number of weeks
+        var difference = endDate - startDate;
+        return Math.round(difference / (60 * 60 * 24 * 7));
+    };
+
+    /**
+     * Split up a given Array of events into separate Arrays per term. Returns an object with
+     * keys being the term label and values the Array of events associated to that term
+     *
+     * @param  {Event[]}    events    The Array of events to split up into terms
+     * @return {Object}               Object representing the split up terms and events
+     */
+    var splitEventsByTerm = exports.splitEventsByTerm = function(events) {
+        // Get the configuration
+        var config = require('gh.core').config;
+        // Get the correct terms associated to the current application
+        var terms = config.terms[config.academicYear];
+        // Create the object to return
+        var eventsByTerm = {};
+
+        // Transform the start and end dates in the terms to proper Date objects so we can
+        // compare them to the start times of the events. Also add a key to the object to
+        // return in which we will add the triaged events
+        _.each(terms, function(term) {
+            // Convert start and end strings into proper dates for comparison
+            term.start = new Date(term.start);
+            term.end = new Date(term.end);
+            // Add an Array to the object to return with the term label as the key
+            eventsByTerm[term.label] = [];
+        });
+
+        // Loop over the array of events to triage them
+        _.each(events.results, function(ev) {
+            // Convert start date into proper date for comparison
+            var evStart = new Date(ev.start);
+            // Loop over the terms and check whether the event start date
+            // falls between the term start and end date. If it does, push
+            // it into the term's Array
+            _.each(terms, function(term) {
+                if (evStart >= term.start && evStart <= term.end) {
+                    // The event takes place in this term, push it into the Array
+                    eventsByTerm[term.label].push(ev);
+                }
+            });
+        });
+
+        // Return the resulting object
+        return eventsByTerm;
+    };
+});
